BK-10: Rename user module;
