+++ package.json
@@ -21,6 +21,7 @@
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
+    "@nestjs-modules/mailer": "^1.6.1",
     "@nestjs/bullmq": "^10.2.3",
     "@nestjs/class-validator": "^0.13.1",
     "@nestjs/common": "^10.0.0",
@@ -28,6 +29,7 @@
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/typeorm": "^10.0.2",
+    "@types/nodemailer": "^6.4.17",
     "bcrypt": "^5.1.1",
     "bullmq": "^5.34.2",
     "class-transformer": "^0.5.1",
