+++ src/auth/auth.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { UserModule } from 'src/user/user.module';
-import { UserService } from 'src/user/user.service';
+import { UsersModule } from 'src/users/users.module';
+import { UsersService } from 'src/users/users.service';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 
 @Module({
-  imports: [UserModule],
-  providers: [AuthService, UserService],
+  imports: [UsersModule],
+  providers: [AuthService, UsersService],
   controllers: [AuthController],
 })
 export class AuthModule {}
