+++ src/templates/confirmation.hbs
@@ -1,0 +1,3 @@
+<h1>Welcome to Bookly!</h1>
+<p>Please confirm your account by visiting this link:</p>
+<a href="{{confirmationUrl}}">Confirm Email</a>
