+++ src/users/entities/user.entity.ts
@@ -1,6 +1,7 @@
 import * as bcrypt from 'bcrypt';
 import { BasicEntity } from '../../database/entities/basic.entity';
 import { Column, Entity, Index, BeforeInsert } from 'typeorm';
+import { v4 } from 'uuid';
 
 @Entity('users')
 export class UserEntity extends BasicEntity {
@@ -14,9 +15,20 @@ export class UserEntity extends BasicEntity {
   @Column({ type: 'varchar' })
   password: string;
 
+  @Column({ default: false })
+  isConfirmed: boolean;
+
+  @Column({ nullable: true})
+  confirmationToken: string;
+
   @BeforeInsert()
   async encryptPassword() {
     this.password = await bcrypt.hash(this.password, 10);
+  }
+
+  @BeforeInsert()
+  async generateConfirmationToken() {
+    this.confirmationToken = await v4();
   }
 
   @BeforeInsert()
